backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.1
pydantic[email]==2.10.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI / ML
anthropic==0.40.0
chromadb==0.5.20
sentence-transformers==3.3.1
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.3

# HTTP / News
httpx==0.28.0

# Utils
python-dotenv==1.0.1
python-multipart==0.0.17
twilio==9.4.0
